refactor: move client code into subpackage
