+++ internal/provider/client/synthetic_check.go
@@ -1,4 +1,4 @@
-package provider
+package client
 
 import (
 	"context"
@@ -12,9 +12,9 @@ import (
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 )
 
-func (c *dash0Client) CreateView(ctx context.Context, check model.ViewResourceModel) error {
+func (c *dash0Client) CreateSyntheticCheck(ctx context.Context, check model.SyntheticCheckResourceModel) error {
 	// Build URL with dataset query parameter
-	apiPath := fmt.Sprintf("/api/views/%s", check.Origin.ValueString())
+	apiPath := fmt.Sprintf("/api/synthetic-checks/%s", check.Origin.ValueString())
 	u, err := url.Parse(apiPath)
 	if err != nil {
 		return fmt.Errorf("error parsing API path: %w", err)
@@ -26,12 +26,12 @@ func (c *dash0Client) CreateView(ctx context.Context, check model.ViewResourceMo
 	u.RawQuery = q.Encode()
 
 	// Convert YAML to JSON
-	jsonBody, err := converter.ConvertYAMLToJSON(check.ViewYaml.ValueString())
+	jsonBody, err := converter.ConvertYAMLToJSON(check.SyntheticCheckYaml.ValueString())
 	if err != nil {
-		return fmt.Errorf("error converting view YAML to JSON: %w", err)
+		return fmt.Errorf("error converting synthetic check YAML to JSON: %w", err)
 	}
 
-	tflog.Debug(ctx, fmt.Sprintf("Creating view with JSON payload: %s", jsonBody))
+	tflog.Debug(ctx, fmt.Sprintf("Creating synthetic check with JSON payload: %s", jsonBody))
 
 	// Make the API request with JSON
 	resp, err := c.doRequest(ctx, http.MethodPut, u.String(), jsonBody)
@@ -39,12 +39,13 @@ func (c *dash0Client) CreateView(ctx context.Context, check model.ViewResourceMo
 		return err
 	}
 
-	tflog.Debug(ctx, fmt.Sprintf("view created. Got API response: %s", resp))
+	tflog.Debug(ctx, fmt.Sprintf("Synthetic check created. Got API response: %s", resp))
 
 	return nil
 }
-func (c *dash0Client) GetView(ctx context.Context, dataset string, origin string) (*model.ViewResourceModel, error) {
-	apiPath := fmt.Sprintf("/api/views/%s", origin)
+
+func (c *dash0Client) GetSyntheticCheck(ctx context.Context, dataset string, origin string) (*model.SyntheticCheckResourceModel, error) {
+	apiPath := fmt.Sprintf("/api/synthetic-checks/%s", origin)
 	u, err := url.Parse(apiPath)
 	if err != nil {
 		return nil, fmt.Errorf("error parsing API path: %w", err)
@@ -60,18 +61,19 @@ func (c *dash0Client) GetView(ctx context.Context, dataset string, origin string
 		return nil, err
 	}
 
-	check := &model.ViewResourceModel{
-		Origin:   types.StringValue(origin),
-		Dataset:  types.StringValue(dataset),
-		ViewYaml: types.StringValue(string(resp)),
+	check := &model.SyntheticCheckResourceModel{
+		Origin:             types.StringValue(origin),
+		Dataset:            types.StringValue(dataset),
+		SyntheticCheckYaml: types.StringValue(string(resp)),
 	}
 	return check, nil
 }
-func (c *dash0Client) UpdateView(ctx context.Context, check model.ViewResourceModel) error {
+
+func (c *dash0Client) UpdateSyntheticCheck(ctx context.Context, check model.SyntheticCheckResourceModel) error {
 	dataset := check.Dataset.ValueString()
 
 	// Build URL with dataset query parameter
-	apiPath := fmt.Sprintf("/api/views/%s", check.Origin.ValueString())
+	apiPath := fmt.Sprintf("/api/synthetic-checks/%s", check.Origin.ValueString())
 	u, err := url.Parse(apiPath)
 	if err != nil {
 		return fmt.Errorf("error parsing API path: %w", err)
@@ -82,15 +84,15 @@ func (c *dash0Client) UpdateView(ctx context.Context, check model.ViewResourceMo
 	q.Set("dataset", dataset)
 	u.RawQuery = q.Encode()
 
-	tflog.Debug(ctx, fmt.Sprintf("Updating view in dataset: %s", dataset))
+	tflog.Debug(ctx, fmt.Sprintf("Updating synthetic check in dataset: %s", dataset))
 
 	// Convert YAML to JSON
-	jsonBody, err := converter.ConvertYAMLToJSON(check.ViewYaml.ValueString())
+	jsonBody, err := converter.ConvertYAMLToJSON(check.SyntheticCheckYaml.ValueString())
 	if err != nil {
-		return fmt.Errorf("error converting view YAML to JSON: %w", err)
+		return fmt.Errorf("error converting synthetic check YAML to JSON: %w", err)
 	}
 
-	tflog.Debug(ctx, fmt.Sprintf("Updating view with JSON payload: %s", jsonBody))
+	tflog.Debug(ctx, fmt.Sprintf("Updating synthetic check with JSON payload: %s", jsonBody))
 
 	// Make the API request with JSON
 	_, err = c.doRequest(ctx, http.MethodPut, u.String(), jsonBody)
@@ -98,13 +100,14 @@ func (c *dash0Client) UpdateView(ctx context.Context, check model.ViewResourceMo
 		return err
 	}
 
-	tflog.Debug(ctx, fmt.Sprintf("view updated with origin: %s", check.Origin))
+	tflog.Debug(ctx, fmt.Sprintf("Synthetic check updated with origin: %s", check.Origin))
 
 	return nil
 }
-func (c *dash0Client) DeleteView(ctx context.Context, origin string, dataset string) error {
+
+func (c *dash0Client) DeleteSyntheticCheck(ctx context.Context, origin string, dataset string) error {
 	// Build URL with dataset query parameter
-	apiPath := fmt.Sprintf("/api/views/%s", origin)
+	apiPath := fmt.Sprintf("/api/synthetic-checks/%s", origin)
 	u, err := url.Parse(apiPath)
 	if err != nil {
 		return fmt.Errorf("error parsing API path: %w", err)
@@ -115,7 +118,7 @@ func (c *dash0Client) DeleteView(ctx context.Context, origin string, dataset str
 	q.Set("dataset", dataset)
 	u.RawQuery = q.Encode()
 
-	tflog.Debug(ctx, fmt.Sprintf("Deleting view in dataset: %s", dataset))
+	tflog.Debug(ctx, fmt.Sprintf("Deleting synthetic check in dataset: %s", dataset))
 
 	// Make the API request
 	_, err = c.doRequest(ctx, http.MethodDelete, u.String(), "")
@@ -123,7 +126,7 @@ func (c *dash0Client) DeleteView(ctx context.Context, origin string, dataset str
 		return err
 	}
 
-	tflog.Debug(ctx, fmt.Sprintf("view deleted with origin: %s", origin))
+	tflog.Debug(ctx, fmt.Sprintf("Synthetic check deleted with origin: %s", origin))
 
 	return nil
 }
