+++ internal/provider/dashboard_resource_read_test.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"testing"
 
+	"github.com/dash0/terraform-provider-dash0/internal/provider/client"
 	"github.com/dash0/terraform-provider-dash0/internal/provider/model"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
@@ -15,7 +16,7 @@ import (
 
 // Custom mock client implementation for this test
 type testDashboardClient struct {
-	dash0ClientInterface
+	client.Client
 	getResponse *model.DashboardResourceModel
 	getError    error
 }
