+++ internal/provider/synthetic_check_resource_acc_test.go
@@ -6,6 +6,7 @@ import (
 	"os"
 	"testing"
 
+	"github.com/dash0/terraform-provider-dash0/internal/provider/client"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
 )
@@ -209,7 +210,7 @@ func testAccCheckSyntheticCheckExists(resourceName string) resource.TestCheckFun
 		dataset := rs.Primary.Attributes["dataset"]
 
 		// Create a new client to verify the synthetic check exists
-		client := newDash0Client(
+		client := client.NewDash0Client(
 			os.Getenv("DASH0_URL"),
 			os.Getenv("DASH0_AUTH_TOKEN"),
 		)
